+++ hindent.cabal
@@ -40,9 +40,9 @@ library
       HIndent.Ast.Declaration
       HIndent.Ast.Declaration.Collection
       HIndent.Ast.Declaration.Family.Data
-      HIndent.Ast.Declaration.Family.Data.Injectivity
-      HIndent.Ast.Declaration.Family.ResultSignature
       HIndent.Ast.Declaration.Family.Type
+      HIndent.Ast.Declaration.Family.Type.Injectivity
+      HIndent.Ast.Declaration.Family.Type.ResultSignature
       HIndent.Ast.FileHeaderPragma
       HIndent.Ast.FileHeaderPragma.Collection
       HIndent.Ast.Import
